FIX: server cad

+++ server/include/component/enemy/Enemy.hpp
@@ -26,15 +26,15 @@ enum class EnemyType {
 
 class Enemy {
   public:
-    Enemy(EnemyType type, const Point& position, int16_t speed, int16_t width,
-          int16_t height, int16_t bulletSpeed, int16_t bulletDamage,
+    Enemy(EnemyType type, const Point& position, float speed, int16_t width,
+          int16_t height, float bulletSpeed, int16_t bulletDamage,
           int16_t shootCooldown, int16_t shootRange, int16_t health, bool isAlive = true);
 
     int32_t getId() const;
     EnemyType getType() const;
     const Point& getPosition() const;
-    int16_t getSpeed() const;
-    int16_t getBulletSpeed() const;
+    float getSpeed() const;
+    float getBulletSpeed() const;
     int16_t getBulletDamage() const;
     int16_t getWidth() const;
     int16_t getHeight() const;
@@ -56,10 +56,10 @@ class Enemy {
     int32_t _id;
     EnemyType _type;
     Point _position;
-    int16_t _speed;
+    float _speed;
     int16_t _width;
     int16_t _height;
-    int16_t _bulletSpeed;
+    float _bulletSpeed;
     int16_t _bulletDamage;
     int16_t _shootCooldown;
     int16_t _shootRange;
